♻️ Flytt tester fra egen fil inn i `BoutgifterRevuderingIntergraionTest`
